requirements.txt
exa-py>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.0.0
apify-client>=1.0.0
httpx>=0.27.0
reportlab>=4.0.0
